requirements.txt
# Enhanced Save Restricted Content Bot v3 - Dependencies
telethon
python-dotenv
psutil
opencv-python-headless
devgagantools
aiofiles
# Enhanced Pyrogram (custom build)
https://www.dl.dropboxusercontent.com/scl/fi/e0fo6fcjn8kmr5r0x6wvg/myownpyro.zip?rlkey=d1znpwckss4ullz0sg7e1qjjg&st=kmbh7wdv&dl=0
aiohttp
telethon-tgcrypto
motor
pymongo
pytz
Pillow
tgcrypto
flask
werkzeug==2.2.2
mutagen
yt-dlp
requests
cryptography

# Enhanced features dependencies
asyncio-mqtt
httpx
ujson
orjson
lxml
beautifulsoup4
python-magic
ffmpeg-python